+++ cpskin/core/interfaces.py
@@ -165,6 +165,13 @@ class ICPSkinSettings(Interface):
         default=False
     )
 
+    show_footer_sitemap = schema.Bool(
+        title=_(u'Show footer sitemap'),
+        description=_(u'Automatically generate sitemap footer'),
+        required=False,
+        default=True
+    )
+
 
 class IVideoCollection(Interface):
     """